Mean Structure Vector \(\mathbf{m}\)

Derives the mean structure vector \(\mathbf{m}\) using the Reticular Action Model (RAM) notation.

m(mutheta, A, filter = NULL)

Arguments

mutheta
t x 1 numeric vector \(\boldsymbol{\mu} \left( \boldsymbol{\theta} \right)_{t \times 1}\) . Model-implied mean vector.

A
t x t numeric matrix \(\mathbf{A}_{t \times t}\). Asymmetric paths (single-headed arrows), such as regression coefficients and factor loadings.

filter
j x t numeric matrix \(\mathbf{F}_{j \times t}\). Filter matrix used to select variables. If filter = NULL, the filter matrix \(\mathbf{F}\) is assumed to be a \(t \times t\) identity matrix.

Value

Returns the mean structure vector \(\mathbf{m}\) derived from the mutheta, A, and filter matrices.

Details

The mean structure vector \(\mathbf{m}\) as a function of Reticular Action Model (RAM) matrices is given by

$$ \mathbf{m} = \left[ \mathbf{F} \left( \mathbf{I} - \mathbf{A} \right)^{-1} \right]^{-1} \boldsymbol{\mu} \left( \boldsymbol{\theta} \right) $$

where

\(\mathbf{A}_{t \times t}\) represents asymmetric paths (single-headed arrows), such as regression coefficients and factor loadings,

\(\mathbf{F}_{j \times t}\) represents the filter matrix used to select the observed variables,

\(\mathbf{I}_{t \times t}\) represents an identity matrix,

\(\boldsymbol{\mu} \left( \boldsymbol{\theta} \right)\) is the \(t \times 1\) model-implied mean vector

\(j\) number of observed variables,

\(k\) number of latent variables, and

\(t\) number of observed and latent variables, that is \(j + k\) .

References

McArdle, J. J. (2005). The development of the RAM rules for latent variable structural equation modeling. In A. Maydeu-Olivares & J. J. McArdle (Eds.), Contemporary Psychometrics: A festschrift for Roderick P. McDonald (pp. 225--273). Lawrence Erlbaum Associates.

McArdle, J. J., & McDonald, R. P. (1984). Some algebraic properties of the Reticular Action Model for moment structures. British Journal of Mathematical and Statistical Psychology, 37 (2), 234--251.

See also

Author

Ivan Jacob Agaloos Pesigan